+++ src/annbatch/samplers/__init__.py
@@ -1,11 +1,11 @@
 from ._chunk_sampler import ChunkSampler
-from ._distributed_random_sampler import DistributedRandomSampler
+from ._distributed_sampler import DistributedSampler
 from ._random_sampler import RandomSampler
 from ._sequential_sampler import SequentialSampler
 
 __all__ = [
     "ChunkSampler",
-    "DistributedRandomSampler",
+    "DistributedSampler",
     "RandomSampler",
     "SequentialSampler",
 ]